+++ api/src/utils/github.ts
@@ -1,6 +1,7 @@
 import A2A from 'a2a';
 import { graphql } from '@octokit/graphql';
 import dotenv from 'dotenv';
+import { References } from './bundle';
 dotenv.config();
 
 const getGitHubToken = (() => {
@@ -55,6 +56,9 @@ type PageContentsQuery = {
     mdxIndex?: {
       text: string;
     };
+    referenceConfig?: {
+      text: string;
+    };
   };
 };
 
@@ -69,6 +73,7 @@ export type Contents = {
   md: string | null;
   path: string;
   repositoryFound: boolean;
+  referenceConfig: References | null;
 };
 
 export async function getGitHubContents(metadata: MetaData, noDir?: boolean): Promise<Contents> {
@@ -81,7 +86,7 @@ export async function getGitHubContents(metadata: MetaData, noDir?: boolean): Pr
   const [error, response] = await A2A<PageContentsQuery>(
     getGithubGQLClient()({
       query: `
-      query RepositoryConfig($owner: String!, $repository: String!, $configJson: String!, $configYaml: String!, $configToml: String!, $mdx: String!, $mdxIndex: String!) {
+      query RepositoryConfig($owner: String!, $repository: String!, $configJson: String!, $configYaml: String!, $configToml: String!, $mdx: String!, $mdxIndex: String!, $referenceConfig: String!) {
         repository(owner: $owner, name: $repository) {
           baseBranch: defaultBranchRef {
             name
@@ -112,6 +117,11 @@ export async function getGitHubContents(metadata: MetaData, noDir?: boolean): Pr
               text
             }
           }
+          referenceConfig: object(expression: $referenceConfig) {
+            ... on Blob {
+              text
+            }
+          }
         }
       }
     `,
@@ -122,15 +132,28 @@ export async function getGitHubContents(metadata: MetaData, noDir?: boolean): Pr
       configToml: `${ref}:docs.toml`,
       mdx: `${ref}:${absolutePath}.mdx`,
       mdxIndex: `${ref}:${indexPath}.mdx`,
+      referenceConfig: `${ref}:docs.refs.json`
     }),
   );
 
   // if an error is thrown then the repo is not found, if the repo is private then response = { repository: null }
   if (error || response?.repository === null) {
+    console.error(error.response.errors);
+
     //@ts-ignore
     return {
       repositoryFound: false,
     };
+  }
+
+  let referenceConfig = null;
+  try {
+    referenceConfig = response?.repository.referenceConfig?.text
+      ? JSON.parse(response?.repository.referenceConfig?.text)
+      : null;
+  } catch (e) {
+    console.error('Could not parse reference config');
+    console.error(e);
   }
 
   return {
@@ -144,6 +167,7 @@ export async function getGitHubContents(metadata: MetaData, noDir?: boolean): Pr
     },
     md: response?.repository.mdxIndex?.text || response?.repository.mdx?.text || null,
     path: response?.repository.mdxIndex?.text ? indexPath : absolutePath,
+    referenceConfig,
   };
 }