+++ packages/dart_cli/example/main.dart
@@ -1,0 +1,7 @@
+import 'package:docs_page/src/typedoc.dart';
+
+void main() async {
+  final rootAst = await getTypedocJson();
+
+  await generate(ast: rootAst);
+}