+++ packages/dart_cli/lib/src/help.dart
@@ -1,0 +1,15 @@
+displayHelp() {
+  print('''
+docs_page CLI tool
+
+These are the supported commands:
+
+  init      Initializes the current directory as a Docs.page project.
+
+  typedoc   Generates reference documentation from your docs/typedoc.json.
+  
+  Make sure to complete configuration (see https://use.docs.page/typedoc) to see these in your documentaion.
+
+  Encounter a problem? Raise an issue in our GitHub repository - https://github.com/invertase/docs.page.
+''');
+}