+++ packages/dart_cli/CHANGELOG.md
@@ -1,3 +1,11 @@
-## 1.0.0
+## 0.0.1
 
-- Initial version.
+- prerelease version.
+
+## 0.0.2
+
+- added `help` command and made `init` command more user-friendly
+
+## 0.0.3
+
+- fix check for `docs.refs.json` on `typedoc` command